part4 done

+++ controllers/blogs.js
@@ -2,7 +2,7 @@
 const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+//const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
  
@@ -13,14 +13,17 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  //const user = request.user
+  const user = request.user
   //const token = getTokenFrom(request)
-  const token = request.token
-  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if(!request.token){
+    return response.status(401).json({ error: 'token missing' })
+  }
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
-  const user = await User.findById(decodedToken.id)
+  //const user = await User.findById(decodedToken.id)
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -46,13 +49,11 @@ blogsRouter.post('/', async (request, response) => {
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
 
-  response.json(savedBlog)
+  response.status(201).json(savedBlog)
   
 }) 
 
 blogsRouter.delete('/:id', async (request, response) => {
-  //const token = request.token
-  //const decodedToken = jwt.verify(token, process.env.SECRET)
   const token = request.token
   const decodedToken = jwt.verify(token, process.env.SECRET)
   if (!decodedToken.id) {
@@ -62,11 +63,14 @@ blogsRouter.delete('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
   //const user = await User.findById(decodedToken.id)
   const user = request.user
-  //console.log(user)
+
   if (!blog) {
     return response.status(401).json({ error: 'blog not found' })
   }
 
+  if(!user){
+    return response.status(401).json({ error: 'user not found' })
+  }
 
   if (user.id.toString() === blog.user.toString()) {
     await Blog.findByIdAndRemove(request.params.id)
